--- code/system.ui/react.video/src/index.TEST.mts
@@ -1,7 +1,0 @@
-import { describe, expect, it, Pkg } from './test';
-
-describe('main', () => {
-  it('temp 🐷', () => {
-    expect(Pkg.version.length).to.greaterThan(0);
-  });
-});
