+++ code/system.ui/react.video/src/test/index.ts
@@ -1,2 +1,2 @@
 export { describe, it } from 'vitest';
-export * from '../ui.test';
+export * from '../test.ui';
