+++ code/system.ui/react.video/src/ui/Vimeo/components/ThumbnailOverlay.tsx
